+++ src/spatial_graph/_graph/spatial_graph.py
@@ -4,9 +4,10 @@ from typing import TYPE_CHECKING, Any, ClassVar
 
 import numpy as np
 
-from ._dtypes import DType
-from ._graph.graph import DiGraph, Graph, GraphBase
-from ._rtree import LineRTree, PointRTree
+from spatial_graph._dtypes import DType
+from spatial_graph._rtree import LineRTree, PointRTree
+
+from .graph import DiGraph, Graph, GraphBase
 
 if TYPE_CHECKING:
     from collections.abc import Mapping
