--- tests/test_assert.py
@@ -1,2 +1,0 @@
-def test_assert():
-    assert True
